All sensors can now read data and store to the db, both yocto and phidget
Finished the PhidgetsSensor.java which reads data from the phidget
sensors. Now all five sensors can read and store data to the database.

How-to:

Yocto: Create an instance of YoctopuceSensors.java and use the start()
method to start the thread which reads and stores data from three
sensors.
Phidget: Create an instance of PhidgetsSensor.java and it will start
the event-handled two sensors.

The instantiation should be done in Server.java